boa/pyproject.toml
[project]
name = "boa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "titanoboa>=0.2.7",
]
